feat: v0.5.1 — upgrade to eframe/egui 0.34, fix TUI overlay sizing and cursor
Upgrade all GUI dependencies to latest versions and fix critical TUI mode bugs:

- Upgrade eframe/egui from 0.24 to 0.34 with full API migration
- Upgrade arboard from 3.4 to 3.6
- Upgrade windows-sys from 0.52 to 0.61
- Fix TUI overlay only rendering at ~1/4 window size (PTY never resized)
- Add PTY resize capability via preserved MasterPty handle
- Fix missing cursor in TUI apps like vim
- Fix full-width row rendering in terminal screen buffer

Made-with: Cursor

+++ src/ui/mod.rs
@@ -314,7 +314,7 @@ impl LayoutManager {
             self.responsive_spacing().x * 0.75,
             self.responsive_spacing().y * 0.5,
         );
-        style.spacing.menu_margin = egui::Margin::same(self.responsive_spacing().x * 0.5);
+        style.spacing.menu_margin = egui::Margin::same((self.responsive_spacing().x * 0.5) as i8);
 
         // Update font sizes
         style.text_styles.insert(
